+++ include/Ace/UIMaterial.h
@@ -69,7 +69,7 @@ namespace ace
 			"uniform lowp vec4 Color;										\n"
 			"void main()													\n"
 			"{																\n"
-			"	gl_FragColor = texture2D( Diffuse, o_uv.st) * o_c;			\n"
+			"	gl_FragColor = texture2D( Diffuse, o_uv.st) * o_c * Color;	\n"
 			"}																\n"
 			, ShaderType::Fragment);